pyproject.toml
[project]
name = "langsmith-scape-eval"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.26",
    "langchain-openai>=0.3.27",
    "langsmith>=0.4.4",
    "python-dotenv>=1.1.1",
]
